rename file before adding activity
