+++ data/models/ingredient.py
@@ -1,10 +1,12 @@
 from django.db import models
+
 from simple_history.models import HistoricalRecords
 
-from data.behaviours import TimeStampable, Historisable
-from .mixins import WithMissingImportBoolean, WithComments
-from .status import WithStatus
+from data.behaviours import Historisable, TimeStampable
+
 from .abstract_models import CommonModel
+from .ingredient_status import WithStatus
+from .mixins import WithComments, WithMissingImportBoolean
 from .substance import Substance
